+++ qgreenland/config/layers/Biology/Birds/naturemap.py
@@ -1,0 +1,3 @@
+from qgreenland.config.helpers.layers.naturemap import make_birds_layers
+
+layers = make_birds_layers()