+++ qgreenland/config/layers/Biology/Mammals/naturemap.py
@@ -1,0 +1,3 @@
+from qgreenland.config.helpers.layers.naturemap import make_mammals_layers
+
+layers = make_mammals_layers()